pyproject.toml
[tool.poetry]
name = "digital_CO2_footprint"
version = "0.1.0"
description = "Calculate digital video streaming CO2 footprint with a Streamlit UI"
authors = ["Stéphane Debove"]
readme = "README.md"
packages = [{ include = "*", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.12,<4.0.0"
loguru = ">=0.7.3,<0.8.0"
streamlit = ">=1.38.0,<2.0.0"
pyyaml = ">=6.0.2,<7.0.0"

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
